--- src/main/scala/experimental/TryAgain.scala
@@ -1,373 +1,0 @@
-package experimental
-
-import matrix.Matrix
-
-import java.math.{MathContext, RoundingMode}
-import scala.annotation.tailrec
-import org.apache.commons.math3.fraction.Fraction
-
-type A = Fraction
-val Zero: A = Fraction.ZERO
-val One: A  = Fraction.ONE
-
-extension (a: A) {
-  def /(b: A): A = a.divide(b)
-  def -(b: A): A = a.subtract(b)
-  def *(b: A): A = a.multiply(b)
-}
-
-type MyVector[B] = List[B]
-
-import scala.annotation.tailrec
-
-extension [A](list: List[A]) {
-  def traverse[B](f: A => Option[B]): Option[List[B]] = {
-    @tailrec
-    def traverseHelper(remaining: List[A], acc: List[B]): Option[List[B]] = {
-      remaining match {
-        case Nil => Some(acc.reverse)
-        case head :: tail =>
-          f(head) match {
-            case Some(b) => traverseHelper(tail, b :: acc)
-            case None    => None
-          }
-      }
-    }
-
-    traverseHelper(list, Nil)
-  }
-}
-
-def combineOptions[A, B, C](optionA: Option[A], optionB: Option[B])(func: (A, B) => C): Option[C] =
-  optionA.flatMap(a => optionB.map(b => func(a, b)))
-
-def map2(a: MyVector[A], b: MyVector[A])(f: (A, A) => A): MyVector[A] =
-  a.zip(b).map { case (aX, bX) => f(aX, bX) }
-
-def lmap2[A](a: List[A], b: List[A])(f: (A, A) => A): List[A] =
-  a.zip(b).map { case (aX, bX) => f(aX, bX) }
-
-def desplit[A](lead: A, tail: Option[MyVector[A]]): MyVector[A] =
-  lead :: tail.getOrElse(Nil)
-
-def split[A](vector: MyVector[A]): (A, Option[MyVector[A]]) =
-  vector match
-    case element :: Nil        => element -> None
-    case element :: tailVector => element -> Some(tailVector)
-
-@tailrec
-def moveNonZeroLeadRowToTheTopInternal(
-  matrixLeft: MyVector[MyVector[A]],
-  skippedRows: MyVector[MyVector[A]],
-): MyVector[MyVector[A]] =
-  val (headVector, maybeTailVectors) = split(matrixLeft)
-  if headVector.head == Zero then
-    maybeTailVectors match
-      case Some(tailVectors) => moveNonZeroLeadRowToTheTopInternal(tailVectors, headVector :: skippedRows)
-      case None              => headVector :: skippedRows
-  else headVector :: maybeTailVectors.fold(skippedRows) { tailVectors => skippedRows ++ tailVectors }
-
-def moveNonZeroLeadRowToTheTop(matrix: MyVector[MyVector[A]]): MyVector[MyVector[A]] =
-  val (headVector, maybeTailVectors) = split(matrix)
-  if headVector.head != Zero then matrix
-  else
-    maybeTailVectors.fold(matrix) { tailVectors =>
-      moveNonZeroLeadRowToTheTopInternal(tailVectors, List(headVector))
-    }
-
-def sliceRightForSubtract(
-  vector: MyVector[A],
-  partToSubtract: Option[MyVector[A]],
-): (Option[MyVector[A]], A, Option[MyVector[A]]) =
-  partToSubtract match
-    case None =>
-      val localLead = vector.takeRight(1).head
-      val maybeRest = vector.dropRight(1)
-      (Option.when(maybeRest.nonEmpty)(maybeRest), localLead, None)
-
-    case Some(vectorToSubtract) =>
-      val localLead :: subtractionPart = vector.takeRight(vectorToSubtract.length + 1)
-      val maybeRest                    = vector.dropRight(vectorToSubtract.length + 1)
-      (Option.when(maybeRest.nonEmpty)(maybeRest), localLead, Some(subtractionPart))
-
-def printMatrix(matrix: List[List[A]]): Unit = {
-  // Get the number of rows and columns in the matrix
-  val numRows = matrix.length
-  val numCols = if (numRows > 0) matrix.head.length else 0
-
-  // Find the maximum length of the elements in the matrix
-  val maxElementLength = matrix.flatten.map(_.toString.length).max
-
-  // Print the matrix
-  for (row <- matrix) {
-    for (element <- row) {
-      val paddedElement = element.toString.padTo(maxElementLength, ' ')
-      print(s"$paddedElement ")
-    }
-    println()
-  }
-}
-
-def moveNonZeroLeadRowToTop2(matrix: MyVector[MyVector[A]]): MyVector[MyVector[A]] = {
-  // todo: do maybe slices and finding the index to swap
-  moveNonZeroLeadRowToTheTop(matrix)
-}
-
-type NonEmptyList[+A] = List[A]
-
-case class SmallerResult(
-  smallerMatrix: MyVector[MyVector[A]],
-  toBeSubtractedAbove: List[NonEmptyList[Option[MyVector[A]]]],
-)
-
-def onZeroColumn(height: Int, maybeRightMatrixToProcess: Option[MyVector[MyVector[A]]]): SmallerResult =
-  maybeRightMatrixToProcess match
-    case Some(rightMatrixToProcess) =>
-      val SmallerResult(rightMatrix, toBeSubtractedAbove) = reducedRecursive(rightMatrixToProcess)
-      SmallerResult(rightMatrix.map(Zero :: _), toBeSubtractedAbove)
-    case None =>
-      SmallerResult(List.tabulate(height)(_ => List(Zero)), List.empty)
-
-def dropZeroColumn(
-  headVectorTail: MyVector[A],
-  maybeTailMatrix: Option[MyVector[MyVector[A]]],
-): MyVector[MyVector[A]] =
-  desplit(
-    headVectorTail,
-    maybeTailMatrix.map(_.map(_.tail)),
-  ) // todo: it should be guaranteed, that if we have headVectorTail we can do _.tail operation
-
-def subtractBy(coefficient: A)(x: A, base: A): A =
-  x - base * coefficient
-
-def subtractDown(
-  headVectorLead: A,
-  maybeHeadVectorTail: Option[MyVector[A]],
-  maybeTailMatrix: Option[MyVector[MyVector[A]]],
-): Option[MyVector[MyVector[A]]] =
-  maybeTailMatrix.flatMap { tailMatrix =>
-    tailMatrix.traverse { tailMatrixRow =>
-      val (rowLead, maybeRowTail) = split(tailMatrixRow)
-      if rowLead == Zero then maybeRowTail
-      else
-        combineOptions(maybeHeadVectorTail, maybeRowTail) { (headVectorTail, rowTail) =>
-          map2(rowTail, headVectorTail)(subtractBy(rowLead / headVectorLead))
-        }
-    }
-  }
-
-@tailrec
-def subtractBack(
-  toSubtract: List[NonEmptyList[Option[MyVector[A]]]],
-  vector: MyVector[A],
-  vectorParts: List[Option[MyVector[A]]] = List.empty,
-): NonEmptyList[Option[MyVector[A]]] = {
-  toSubtract match
-    case Nil => Some(vector) :: vectorParts
-    case currentToSubtract :: othersToSubtract =>
-      val splitPattern                                   = currentToSubtract.head
-      val (maybeRest, localLead, maybeNewPartToSubtract) = sliceRightForSubtract(vector, splitPattern)
-
-      val subtractedParts = lmap2[Option[MyVector[A]]](maybeNewPartToSubtract :: vectorParts, currentToSubtract) {
-        (maybeVectorPart, maybeToSubtractPart) =>
-          combineOptions(maybeVectorPart, maybeToSubtractPart) { (vectorPart, toSubtractPart) =>
-            map2(vectorPart, toSubtractPart)(subtractBy(localLead))
-          }
-      }
-
-      maybeRest match
-        case Some(restVector) =>
-          subtractBack(
-            othersToSubtract,
-            restVector,
-            subtractedParts,
-          )
-        case None => None :: subtractedParts
-}
-
-def divideByLead(lead: A, maybeVector: Option[MyVector[A]]): Option[MyVector[A]] =
-  maybeVector.map(_.map(_ / lead))
-
-def composeVectorParts(parts: NonEmptyList[Option[MyVector[A]]]): Option[MyVector[A]] = {
-  val (_, withoutLeftZero) = split(parts.flatMap(desplit(Zero, _)))
-  withoutLeftZero
-}
-
-def onDownSubtraction(
-  headLead: A,
-  maybeHeadTail: Option[MyVector[A]],
-  maybeDownRightMatrixToProcess: Option[MyVector[MyVector[A]]],
-): SmallerResult =
-  maybeDownRightMatrixToProcess match
-    case Some(downRightMatrixToProcess) =>
-      val SmallerResult(downRightMatrix, toSubtract) = reducedRecursive(downRightMatrixToProcess)
-
-      val maybeSubtractedTailParts: Option[NonEmptyList[Option[MyVector[A]]]] =
-        maybeHeadTail.map(subtractBack(toSubtract, _))
-
-      val maybeDividedSubtractedTailParts: Option[NonEmptyList[Option[MyVector[A]]]] =
-        maybeSubtractedTailParts.map(_.map(divideByLead(headLead, _)))
-
-      val maybeComposedVectorTail: Option[MyVector[A]] =
-        maybeDividedSubtractedTailParts.flatMap(composeVectorParts)
-
-      val headVector       = desplit(One, maybeComposedVectorTail)
-      val zeroedDownMatrix = downRightMatrix.map(Zero :: _)
-      SmallerResult(
-        headVector :: zeroedDownMatrix,
-        maybeDividedSubtractedTailParts.fold(toSubtract)(toSubtract :+ _),
-      )
-    case None =>
-      val maybeDividedTail = divideByLead(headLead, maybeHeadTail)
-      SmallerResult(List(desplit(One, maybeDividedTail)), List(List(maybeDividedTail)))
-
-def reducedRecursive(matrix: MyVector[MyVector[A]]): SmallerResult = {
-  val swapped = moveNonZeroLeadRowToTop2(matrix)
-
-  val (headVector, maybeTailMatrix)         = split(swapped)
-  val (headVectorLead, maybeHeadVectorTail) = split(headVector)
-
-  if headVectorLead == Zero then
-    onZeroColumn(
-      matrix.length,
-      maybeHeadVectorTail.map(dropZeroColumn(_, maybeTailMatrix)),
-    )
-  else
-    onDownSubtraction(
-      headVectorLead,
-      maybeHeadVectorTail,
-      subtractDown(headVectorLead, maybeHeadVectorTail, maybeTailMatrix),
-    )
-}
-
-def reduced(matrix: MyVector[MyVector[A]]): MyVector[MyVector[A]] =
-  reducedRecursive(matrix).smallerMatrix
-
-//
-//
-//
-//
-//
-
-@main def test = {
-  println("start!")
-
-  given Conversion[Int, A] = int => Fraction(int)
-
-  val matrix1: MyVector[MyVector[A]] = List(
-    List(0, 2, 3, 4),
-    List(0, 6, 7, 8),
-    List(0, 3, 2, 1),
-  )
-
-  val matrix2: MyVector[MyVector[A]] = List(
-    List(1, 3, 1, 9),
-    List(1, 3, 5, 9),
-    List(1, 1, -1, 1),
-  )
-
-  val matrix3: MyVector[MyVector[A]] = List(
-    List(1, 2, 3, 4, 5, 6),
-    List(0, 0, 1, 2, 3, 4),
-    List(0, 0, 0, 0, 1, 2),
-  )
-
-  val matrix4: MyVector[MyVector[A]] = List(
-    List(1, 2, 3, 4, 5, 6),
-    List(0, 0, 1, 2, 3, 4),
-    List(0, 0, 0, 0, 0, 0),
-  )
-
-  val matrix5: MyVector[MyVector[A]] = List(
-    List(0, 0, 0, 0),
-    List(0, 0, 0, 0),
-    List(0, 0, 0, 0),
-  )
-
-  val matrix6: MyVector[MyVector[A]] = List(
-    List(1, 2, 3, 4, 5, 6, 7, 8, 9),
-    List(0, 0, 0, 1, 2, 3, 4, 5, 6),
-    List(0, 0, 0, 0, 0, 0, 1, 2, 3),
-  )
-
-  def process(matrix: MyVector[MyVector[A]]): Unit = {
-    println("In:")
-    printMatrix(matrix)
-    println("Out:")
-    printMatrix(reduced(matrix))
-    println("---")
-  }
-
-  val randomMatrix1: MyVector[MyVector[A]] = List(
-    List(4, 19, 0, 1),
-    List(42, 4, -5, 0),
-    List(0, 1, -9, 20),
-  )
-  /* the answer is:
-  1            0            0            -1931 / 7058
-  0            1            0            389 / 3529
-  0            0            1            -7799 / 3529
-   */
-
-  val randomMatrix2: MyVector[MyVector[A]] = List(
-    List(-1, -2, -3, -4, -5, -6),
-    List(-1, -2, 1, 2, 3, 4),
-    List(5, 6, 7, 8, 9, 10),
-    List(-5, -6, -7, -8, -9, -10),
-  )
-  /* the answer is:
-	1	0	0	-0.5	-1	-1.5
-	0	1	0	0	0	0
-	0	0	1	1.5	2	2.5
-	0	0	0	0	0	0
-   */
-
-  val randomMatrix3: MyVector[MyVector[A]] = List(
-    List(2, 4, -2, 6),
-    List(1, 2, -1, 3),
-    List(3, 2, 1, 11),
-  )
-  /* the answer is:
-	[ 1  0  1 |  4 ]
-    [ 0  1 -1 |  -0.5 ]
-    [ 0  0  0 |  0 ]
-   */
-
-  val randomMatrix4: MyVector[MyVector[A]] = List(
-    List(2, 1, 3, 4, 5, 6, 7, 8, 9, 10, 11),
-    List(12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22),
-    List(23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33),
-    List(34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44),
-    List(45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55),
-    List(56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66),
-    List(67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77),
-    List(78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88),
-    List(89, 90, 91, 92, 93, 94, 95, 96, 97, 98, 99),
-    List(100, 101, 102, 103, 104, 105, 106, 107, 108, 109, 110),
-  )
-  /* the answer is:
-    X1	X2	X3	X4	X5	X6	X7	X8	X9	X10	b
-    1	0	0	-1/3	-2/3	-1	-4/3	-5/3	-2	-7/3	-8/3
-    0	1	0	-1/3	-2/3	-1	-4/3	-5/3	-2	-7/3	-8/3
-    0	0	1	5/3	7/3	3	11/3	13/3	5	17/3	19/3
-    0	0	0	0	0	0	0	0	0	0	0
-    0	0	0	0	0	0	0	0	0	0	0
-    0	0	0	0	0	0	0	0	0	0	0
-    0	0	0	0	0	0	0	0	0	0	0
-    0	0	0	0	0	0	0	0	0	0	0
-    0	0	0	0	0	0	0	0	0	0	0
-    0	0	0	0	0	0	0	0	0	0	0
-   */
-
-  process(matrix1)
-  process(matrix2)
-  process(matrix3)
-  process(matrix4)
-  process(matrix5)
-  process(matrix6)
-
-  process(randomMatrix1)
-  process(randomMatrix2)
-  process(randomMatrix3)
-  process(randomMatrix4)
-}
